Move discovery logic
