+++ test/test_helper.rb
@@ -4,3 +4,4 @@ require 'minitest/spec'
 require 'minitest/autorun'
 require 'turn'
 require 'webmock/minitest'
+require 'json'
